added bottom tab nav on home screen
