+++ navigation/AppNavigator.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { useSelector } from "react-redux";
-import { MenuNavigator, AuthNavigator } from './MainTabNavigator';
+import { MenuNavigator } from './MenuDrawerNavigator';
+import { AuthNavigator } from './AuthStackNavigator';
 import StartupScreen from '../screens/StartupScreen';
 
 const AppNavigator = () => {
