+++ package.json
@@ -19,6 +19,7 @@
     "@fortawesome/react-native-fontawesome": "^0.2.0",
     "@react-native-community/datetimepicker": "^2.1.0",
     "@react-native-community/masked-view": "^0.1.10",
+    "@react-navigation/bottom-tabs": "^5.6.1",
     "@react-navigation/core": "^5.0.0-alpha.16",
     "@react-navigation/drawer": "^5.8.2",
     "@react-navigation/native": "^5.5.1",
